Perform locking in a subquery straight away for lock and update managers
