+++ activerecord/lib/arel/nodes/update_statement.rb
@@ -3,7 +3,7 @@
 module Arel # :nodoc: all
   module Nodes
     class UpdateStatement < Arel::Nodes::Node
-      attr_accessor :relation, :wheres, :values, :orders, :limit, :lock, :offset, :key
+      attr_accessor :relation, :wheres, :values, :orders, :limit, :offset, :key
 
       def initialize
         @relation = nil
@@ -11,7 +11,6 @@ module Arel # :nodoc: all
         @values   = []
         @orders   = []
         @limit    = nil
-        @lock     = nil
         @offset   = nil
         @key      = nil
       end
@@ -23,7 +22,7 @@ module Arel # :nodoc: all
       end
 
       def hash
-        [@relation, @wheres, @values, @orders, @limit, @lock, @offset, @key].hash
+        [@relation, @wheres, @values, @orders, @limit, @offset, @key].hash
       end
 
       def eql?(other)
@@ -33,7 +32,6 @@ module Arel # :nodoc: all
           self.values == other.values &&
           self.orders == other.orders &&
           self.limit == other.limit &&
-          self.lock == other.lock &&
           self.offset == other.offset &&
           self.key == other.key
       end
