+++ activerecord/lib/arel/select_manager.rb
@@ -3,7 +3,23 @@
 module Arel # :nodoc: all
   class SelectManager < Arel::TreeManager
     include Arel::Crud
-    include TreeManager::LockMethods
+
+    def lock(locking = Arel.sql("FOR UPDATE"))
+      case locking
+      when true
+        locking = Arel.sql("FOR UPDATE")
+      when Arel::Nodes::SqlLiteral
+      when String
+        locking = Arel.sql locking
+      end
+
+      @ast.lock = Nodes::Lock.new(locking)
+      self
+    end
+
+    def locked
+      @ast.lock
+    end
 
     STRING_OR_SYMBOL_CLASS = [Symbol, String]
